requirements.txt
pygame
pynput
smtplib
threading
calendar 
pyautogui 
email
requests
getpass
mysql.connector
random
mss
ssl
ctypes
